Software-SUNAT/script- CRM-SUNAT/src/requirements.txt
# Requirements para Cloud Run - CRM SUNAT
# FastAPI application con conexión a Cloud SQL

# Web framework para Cloud Run (FastAPI + Uvicorn)
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Core dependencies
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pg8000==1.30.5

# Cloud SQL Connector (recomendado para Cloud Run)
cloud-sql-python-connector[pg8000]==1.7.0

# Date utilities
python-dateutil==2.8.2

# Google Cloud
google-cloud-logging==3.9.0
